Add DebugUI and DebugMonitor to track Player Coordinates, Waypoint class
Initial commit to setup basic Monitor and UI for debug class.
Currently runs a .5 second loop to update two text boxes with
player coordinates.  Hitting Add will create a D3.Point object and
add it to lstCoordinates.

Added simple Waypoint class, storing a point, level area, quest id
and quest step.  These are created and added to the listbox.

Implemented the save button which uses an XMLSerializer to save
the waypoints stored in the list.